+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Major Changes
 
+- Added `TransactNamespace.simulateAssetChangesBundle()` to simulate a list of transactions in sequence and return a list of asset changes.
+- Added `TransactNamespace.simulateExecutionBundle()` to simulate a list of transactions in sequence and return a list of decoded logs and traces.
+
 ### Minor Changes
 
 ## 2.4.3