+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Major Changes
 
+### Minor Changes
+
+## 2.5.0
+
+### Major Changes
+
 - Added `TransactNamespace.simulateAssetChangesBundle()` to simulate a list of transactions in sequence and return a list of asset changes.
 - Added `TransactNamespace.simulateExecutionBundle()` to simulate a list of transactions in sequence and return a list of decoded logs and traces.
 - Added `NftNamespace.getContractMetadataBatch()` to get the metadata associated with multiple NFT contracts in a single request.